+++ Runtime/Stores/Android/GooglePlay/AAR/Models/ExternalBillingProgramClientInternal.cs
@@ -1,5 +1,6 @@
 using System;
 using Uniject;
+using UnityEngine.Purchasing.GoogleBilling.Interfaces;
 using UnityEngine.Purchasing.Models;
 using UnityEngine.Purchasing.Telemetry;
 using UnityEngine.Scripting;
@@ -10,10 +11,20 @@ namespace UnityEngine.Purchasing.GoogleBilling.Models
     /// The internal counterpart to <see cref="ExternalBillingProgramClient"/>
     /// This class is for implementation details that we don't want to expose publicly
     /// </summary>
-    sealed class ExternalBillingProgramClientInternal : BillingClientBase
+    internal class ExternalBillingProgramClientInternal : BillingClientBase, IExternalBillingProgramClientInternal
     {
         // From <a href="https://developer.android.com/reference/com/android/billingclient/api/BillingClient.BillingProgram#EXTERNAL_CONTENT_LINK()">
         const int k_EXTERNAL_CONTENT_LINK = 1;
+
+        const string k_BillingProgramReportingDetailsParamsClassName =
+            "com.android.billingclient.api.BillingProgramReportingDetailsParams";
+        const string k_LaunchExternalLinkParamsClassName =
+            "com.android.billingclient.api.LaunchExternalLinkParams";
+        const string k_AndroidUriClassName = "android.net.Uri";
+
+        static AndroidJavaClass s_BillingProgramReportingDetailsParamsClass;
+        static AndroidJavaClass s_LaunchExternalLinkParamsClass;
+        static AndroidJavaClass s_AndroidUriClass;
 
         [Preserve]
         internal ExternalBillingProgramClientInternal(IUtil util, ITelemetryDiagnostics telemetryDiagnostics) :
@@ -27,7 +38,7 @@ namespace UnityEngine.Purchasing.GoogleBilling.Models
             m_BillingClient = builder.Call<AndroidJavaObject>("build");
         }
 
-        internal void IsBillingProgramAvailableAsync(Action<IGoogleBillingResult> onBillingProgramAvailabilityResponse)
+        public void IsBillingProgramAvailableAsync(Action<IGoogleBillingResult> onBillingProgramAvailabilityResponse)
         {
             m_BillingClient.Call(
                 "isBillingProgramAvailableAsync",
@@ -36,17 +47,19 @@ namespace UnityEngine.Purchasing.GoogleBilling.Models
             );
         }
 
-        internal void LaunchExternalLink(
+        public void LaunchExternalLink(
             string externalLinkUrl,
             Action<IGoogleBillingResult> onLaunchExternalLinkResponse,
             LinkType linkType,
             LaunchMode launchMode
         )
         {
+            using var newLaunchExternalLinkParams = launchExternalLinkParams(externalLinkUrl, linkType, launchMode);
+
             m_BillingClient.Call(
                 "launchExternalLink",
                 UnityActivity.GetCurrentActivity(),
-                launchExternalLinkParams(externalLinkUrl, linkType, launchMode),
+                newLaunchExternalLinkParams,
                 new LaunchExternalLinkResponseListener(onLaunchExternalLinkResponse, m_Util)
             );
         }
@@ -57,30 +70,49 @@ namespace UnityEngine.Purchasing.GoogleBilling.Models
             LaunchMode launchMode
         )
         {
-            var uri = new AndroidJavaClass("android.net.Uri")
+            using var uri = GetAndroidUriClass()
                 .CallStatic<AndroidJavaObject>("parse", externalLinkUrl);
 
-            var builder = new AndroidJavaClass("com.android.billingclient.api.LaunchExternalLinkParams")
-                .CallStatic<AndroidJavaObject>("newBuilder")
-                .Call<AndroidJavaObject>("setBillingProgram", k_EXTERNAL_CONTENT_LINK)
-                .Call<AndroidJavaObject>("setLinkUri", uri)
-                .Call<AndroidJavaObject>("setLinkType", (int)linkType)
-                .Call<AndroidJavaObject>("setLaunchMode", (int)launchMode);
+            using var builder = GetLaunchExternalLinkParamsClass()
+                .CallStatic<AndroidJavaObject>("newBuilder");
+
+            builder.Call<AndroidJavaObject>("setBillingProgram", k_EXTERNAL_CONTENT_LINK).Dispose();
+            builder.Call<AndroidJavaObject>("setLinkUri", uri).Dispose();
+            builder.Call<AndroidJavaObject>("setLinkType", (int)linkType).Dispose();
+            builder.Call<AndroidJavaObject>("setLaunchMode", (int)launchMode).Dispose();
 
             var launchExternalLinkParams = builder.Call<AndroidJavaObject>("build");
             return launchExternalLinkParams;
         }
 
-        internal void CreateBillingProgramReportingDetailsAsync(Action<IGoogleBillingResult, string> onCreateBillingProgramReportingDetailsResponse)
+        public void CreateBillingProgramReportingDetailsAsync(Action<IGoogleBillingResult, string> onCreateBillingProgramReportingDetailsResponse)
         {
-            var paramsClass = new AndroidJavaClass("com.android.billingclient.api.BillingProgramReportingDetailsParams");
-            var builder = paramsClass.CallStatic<AndroidJavaObject>("newBuilder");
+            using var builder = GetBillingProgramReportingDetailsParamsClass()
+                .CallStatic<AndroidJavaObject>("newBuilder");
+            
+            builder.Call<AndroidJavaObject>("setBillingProgram", k_EXTERNAL_CONTENT_LINK).Dispose();
 
-            var reportingDetailsParams = builder
-                .Call<AndroidJavaObject>("setBillingProgram", k_EXTERNAL_CONTENT_LINK)
-                .Call<AndroidJavaObject>("build");
+            using var reportingDetailsParams = builder.Call<AndroidJavaObject>("build");
 
             m_BillingClient.Call("createBillingProgramReportingDetailsAsync", reportingDetailsParams, new BillingProgramReportingDetailsListener(onCreateBillingProgramReportingDetailsResponse, m_Util));
+        }
+
+        static AndroidJavaClass GetBillingProgramReportingDetailsParamsClass()
+        {
+            s_BillingProgramReportingDetailsParamsClass ??= new AndroidJavaClass(k_BillingProgramReportingDetailsParamsClassName);
+            return s_BillingProgramReportingDetailsParamsClass;
+        }
+
+        static AndroidJavaClass GetLaunchExternalLinkParamsClass()
+        {
+            s_LaunchExternalLinkParamsClass ??= new AndroidJavaClass(k_LaunchExternalLinkParamsClassName);
+            return s_LaunchExternalLinkParamsClass;
+        }
+
+        static AndroidJavaClass GetAndroidUriClass()
+        {
+            s_AndroidUriClass ??= new AndroidJavaClass(k_AndroidUriClassName);
+            return s_AndroidUriClass;
         }
     }
 }