+++ Runtime/Stores/Android/GooglePlay/Interfaces/IBillingClientFactory.cs
@@ -1,0 +1,9 @@
+using UnityEngine.Purchasing.GoogleBilling.Interfaces;
+
+namespace UnityEngine.Purchasing
+{
+    internal interface IBillingClientFactory
+    {
+        IExternalBillingProgramClientInternal CreateExternalBillingProgramClient();
+    }
+}